Rename the unknown byte for model objects
This appears to reference a sound ID used when the object is activated (either
directly, or through a link). Needs testing to see if the specified sound ID is
limited to 255, or if the action flags should be a byte and the sound ID an
int32 (i.e. swap types).